Custom script to repeat Find End of Silence and place markers
Useful to split an album exported as a single audio file.

Slightly better than "Detect CD Tracks" because that one can sometimes
place a marker when the music is everything but silent.

Only downside is that it's not clear how to configure the audio
threshold. So far even 0.000001 seems to make no difference for -70 dB.